# openapi.yaml
openapi: 3.0.0
info:
  title: ServiceDesk API
  version: 1.0.0
  description: |
    Enterprise Service Desk platform with multi-tenant support, comprehensive ticket management,
    knowledge base, analytics, and AI-powered features.

    ## Authentication
    This API uses JWT-based authentication with bearer tokens or secure HTTP-only cookies.
    All authenticated endpoints require a valid token obtained through the login endpoint.

    ## Multi-tenancy
    The API supports multi-tenant architecture. Tenant context is resolved via:
    - Subdomain (tenant.servicedesk.com)
    - HTTP headers (x-tenant-id, x-tenant-slug)
    - Cookie (tenant-context)

    ## Rate Limiting
    API endpoints are rate-limited to prevent abuse. Rate limits vary by endpoint type:
    - Auth endpoints: More restrictive
    - General API endpoints: Standard limits

    Rate limit information is included in response headers:
    - X-RateLimit-Limit
    - X-RateLimit-Remaining
    - X-RateLimit-Reset
  contact:
    name: ServiceDesk API Support
    email: support@servicedesk.com
  license:
    name: Proprietary
    url: https://servicedesk.com/license
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://{tenant}.servicedesk.com/api
    description: Production server (multi-tenant)
    variables:
      tenant:
        default: demo
        description: Tenant subdomain slug
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint

    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: HTTP-only secure cookie containing JWT token
...
  schemas:
    Error:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Error message description"
        code:
          type: string
          example: "ERROR_CODE"
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john.doe@example.com"
        role:
          type: string
          enum: [super_admin, tenant_admin, team_manager, admin, agent, user, manager, read_only, api_client]
          example: "agent"
        tenant_id:
          type: integer
          example: 1
        is_active:
          type: boolean
          example: true
        is_email_verified:
          type: boolean
          example: true
        avatar_url:
          type: string
          nullable: true
          example: "https://example.com/avatar.jpg"
        timezone:
          type: string
          example: "America/Sao_Paulo"
        language:
          type: string
          example: "pt-BR"
        created_at:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
...
  responses:
    UnauthorizedError:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Authentication required"
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Acesso negado - permissão insuficiente"
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Erro interno do servidor"
paths:
...
  /auth/verify:
    get:
      tags:
        - Authentication
      summary: Verify token
      description: Verify if the current authentication token is valid
      operationId: verifyToken
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  valid:
                    type: boolean
                    example: true
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # ==========================================
  # TICKET ENDPOINTS
  # ==========================================
...
  /admin/users:
    get:
      tags:
        - Admin
      summary: List all users
      description: Retrieve list of all users in the tenant (admin only)
      operationId: listUsers
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  users:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/User'
                        - type: object
                          properties:
                            tickets_count:
                              type: integer
                              example: 15
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/ServerError'
...